Extract src/inspectorview.ts — the inspector's markup
Second render-module split, on the same boundary as ./usageview and worth
stating as the convention: a `*view.ts` module takes data and returns a string.
It touches no DOM and calls no renderer, so it needs no seam back into main.ts —
which is why neither of these two needed a hook, unlike the logic slices. The
`render*` functions that put the result in the page stay in main.ts until the
bootstrap trim.

Moved: the vital header, the context/cost gauges, the plan, the working set and
its git buttons, the activity timeline, the resource bars, the diff viewer's
hunk rows, and the small shared helpers behind them (toolVerb/toolClass,
verbFor, dwellText, isLongestWaiting, compactWarn). 171 lines. The moved block
diffs byte-identical against HEAD:src/main.ts apart from the `export` keywords.

Two things had to move for it, both because a module may not import main.ts:

- `statusKey` goes to ./types, beside isAgent. It is a one-line discriminant
  over Sess — literally what types.ts's own header says it holds — and the
  sidebar, mini-rail, tray and inspector pill all read it, so it belongs to none
  of them. GLYPH/GCLASS stay in main.ts with the sidebar rows that use them.
- `gitBusy` goes into inspectorview, which owns the buttons it greys out.
  main.ts's runGit sets it through setGitBusy, the state.ts convention. It did
  not need to go to state.ts: nothing outside those buttons reads it.

Verified in the browser: the six builders render 84 nodes for a synthetic
session with the right content in each — tool verb, the auto-compact warning at
82%, the subagent chip, the plan fraction, +12/↑1 on the working set, the
timeline latency and the memory figure — and setGitBusy still greys the git row
through the relocated flag.
